==> 0-Download from StreetLight.ipynb <==
import os
import pandas as pd

#google cloud big query libaries
from google.cloud import bigquery
from google.oauth2 import service_account

key_path = r"C:\Users\bhereth\streetlight-temp-analysis-e2b201d26862.json"

credentials = service_account.Credentials.from_service_account_file(
    key_path, scopes=["https://www.googleapis.com/auth/cloud-platform"],
)

client = bigquery.Client(credentials=credentials, project=credentials.project_id,)

#directories
working_directory = os.getcwd()
data_folder         = os.path.join(working_directory, "data"        )
intermediate_folder = os.path.join(working_directory, "intermediate")
results_folder      = os.path.join(working_directory, "results"     )


# Import Big Query Data

#ease of use
daytype0 = '0: All Days (Mo-Su)'
daytype1 = '1: Weekday (Tu-Th)'
daytype2 = '2: Weekend Day (Sa-Su)'
dataper1 = '1. All year'
dataper2 = '2. Sep-Nov'
dataper3 = '3. Dec-Feb'
dataper4 = '4. Mar-May'
dataper5 = '5. Jun-Aug'
daypart0 = '0: All Day (12am-12am)'
daypart1 = '1: Early AM (12am-6am)'
daypart2 = '2: Peak AM (6am-9am)'
daypart3 = '3: Mid-Day (9am-3pm)'
daypart4 = '4: Peak PM (3pm-6pm)'
daypart5 = '5: Late PM (6pm-12am)'

#zones for interzonal definition
dImportTables = [
                    ['ut-udot-adap-dev.streetlight_data.udot_commercial_all_year_2019'],
                    ['ut-udot-adap-dev.streetlight_data.udot_commercial_fall_2019'    ],
                    ['ut-udot-adap-dev.streetlight_data.udot_commercial_spring_2019'  ],
                    ['ut-udot-adap-dev.streetlight_data.udot_commercial_summer_2019'  ],
                    ['ut-udot-adap-dev.streetlight_data.udot_commercial_winter_2019'  ]
                ]

sJoinTable = "ut-udot-adap-dev.streetlight_data.wfrc_streetlight_taz"

# Create the pandas DataFrame
dfImportTables = pd.DataFrame(dImportTables, columns = ['tableName'])
display(dfImportTables)


#create query of all subareaid to subareaid truck flows using join to taz table with subareaid field
dfTruckDataBySubareaID = pd.DataFrame()

for dit in dImportTables:
    sql = """
    SELECT
        orig_table.SUBAREAID AS origin_subareaid,
        dest_table.SUBAREAID AS destination_subareaid,
        od_table.mode_of_travel AS mode_of_travel,
        od_table.vehicle_weight AS vehicle_weight,
        od_table.day_type AS day_type,
        od_table.day_part AS day_part,
        od_table.data_period AS data_period,
        SUM(od_table.o_d_traffic_sample_trip_counts) AS od_trip_counts_sum,
        SUM(od_table.o_d_traffic_calibrated_trip_volume) AS od_trip_volume_sum,
        COUNT(mode_of_travel) AS num_records
    FROM
        (""" + dit[0] + """ as od_table
            LEFT JOIN
                """ + sJoinTable + """ AS orig_table
                ON od_table.origin_zone_name = orig_table.SL_COTAZID)
                    LEFT JOIN """ + sJoinTable + """ AS dest_table
                    ON od_table.destination_zone_name = dest_table.SL_COTAZID
    GROUP BY
        orig_table.SUBAREAID,
        dest_table.SUBAREAID,
        od_table.mode_of_travel,
        od_table.vehicle_weight,
        od_table.day_type,
        od_table.day_part,
        od_table.data_period
    """
    #display(sql)

    dfQuery = client.query(sql).to_dataframe()
    dfTruckDataBySubareaID = dfTruckDataBySubareaID.append(dfQuery)

display(dfTruckDataBySubareaID)



#read totals without the join to check against
dfTruckDataSumForChecks = pd.DataFrame()

for dit in dImportTables:

    sqlsumforchecks = """ 
    SELECT
        SUM(od_table.o_d_traffic_sample_trip_counts) AS od_trip_counts_sum,
        SUM(od_table.o_d_traffic_calibrated_trip_volume) AS od_trip_volume_sum,
        COUNT(od_table.mode_of_travel) AS num_records
    FROM
        """ + dit[0] + """ AS od_table
    """

    #display(sqlsumforchecks)

    dfQuerySum = client.query(sqlsumforchecks).to_dataframe()
    dfQuerySum['table'] = dit[0]
    dfTruckDataSumForChecks = dfTruckDataSumForChecks.append(dfQuerySum)

dfTruckDataSumForChecks

display(dfTruckDataSumForChecks.sum())
display(dfTruckDataBySubareaID.sum())

# Large District

#create query of all subareaid to subareaid truck flows using join to taz table with subareaid field
dfTruckDataByDISTLRG = pd.DataFrame()

for dit in dImportTables:
    sql = """
    SELECT
        orig_table.DISTLRG AS origin_distlrg,
        dest_table.DISTLRG AS destination_distlrg,
        od_table.mode_of_travel AS mode_of_travel,
        od_table.vehicle_weight AS vehicle_weight,
        od_table.day_type AS day_type,
        od_table.day_part AS day_part,
        od_table.data_period AS data_period,
        SUM(od_table.o_d_traffic_sample_trip_counts) AS od_trip_counts_sum,
        SUM(od_table.o_d_traffic_calibrated_trip_volume) AS od_trip_volume_sum,
        COUNT(mode_of_travel) AS num_records
    FROM
        (""" + dit[0] + """ as od_table
            LEFT JOIN
                """ + sJoinTable + """ AS orig_table
                ON od_table.origin_zone_name = orig_table.SL_COTAZID)
                    LEFT JOIN """ + sJoinTable + """ AS dest_table
                    ON od_table.destination_zone_name = dest_table.SL_COTAZID
    GROUP BY
        orig_table.DISTLRG,
        dest_table.DISTLRG,
        od_table.mode_of_travel,
        od_table.vehicle_weight,
        od_table.day_type,
        od_table.day_part,
        od_table.data_period
    """
    #display(sql)

    dfQuery = client.query(sql).to_dataframe()
    dfTruckDataByDISTLRG = dfTruckDataByDISTLRG.append(dfQuery)

dfTruckDataByDISTLRG

display(dfTruckDataByDISTLRG.sum())
display(dfTruckDataSumForChecks.sum())
display(dfTruckDataBySubareaID.sum())

# Export to CSV

#export to CSV
dfTruckDataBySubareaID.to_csv(os.path.join(intermediate_folder, r"TruckDataBySubAreaID.csv"))
dfTruckDataByDISTLRG.to_csv(os.path.join(intermediate_folder, r"TruckDataByDISTLRG.csv"))
dfTruckDataSumForChecks.to_csv(os.path.join(intermediate_folder, r"TruckDataSumForChecks.csv"))
