==> 1-Process Data.ipynb <==
import os
import pandas as pd

#directories
working_directory = os.getcwd()
data_folder         = os.path.join(working_directory, "data"        )
intermediate_folder = os.path.join(working_directory, "intermediate")
results_folder      = os.path.join(working_directory, "results"     )

#import from CSV
dfTruckDataBySubareaID = pd.read_csv(os.path.join(intermediate_folder, r"TruckDataBySubAreaID.csv"))
dfTruckDataByDISTLRG = pd.read_csv(os.path.join(intermediate_folder, r"TruckDataByDISTLRG.csv"))
dfTruckDataSumForChecks =  pd.read_csv(os.path.join(intermediate_folder, r"TruckDataSumForChecks.csv"))

#check sums to make sure they match
display(dfTruckDataSumForChecks.sum())
display(dfTruckDataByDISTLRG.sum())
display(dfTruckDataBySubareaID.sum())

# Sub Area

#fill na with -1 meaning EXTERNALS (unmatched TAZs) will be given -1 SubArea
dfTruckDataBySubareaID = dfTruckDataBySubareaID.fillna(-1)
dfTruckDataBySubareaID

dfTruckDataBySubareaID_Pivot_NumRecords = dfTruckDataBySubareaID.pivot_table(values="num_records",index=['day_type', 'day_part', 'data_period', 'origin_subareaid','vehicle_weight'], columns="destination_subareaid",)
dfTruckDataBySubareaID_Pivot_NumRecords = dfTruckDataBySubareaID_Pivot_NumRecords.fillna(0)
dfTruckDataBySubareaID_Pivot_NumRecords = dfTruckDataBySubareaID_Pivot_NumRecords.reset_index()
dfTruckDataBySubareaID_Pivot_NumRecords

dfTruckDataBySubareaID_Pivot_ODTripVolume = dfTruckDataBySubareaID.pivot_table(values="od_trip_volume_sum",index=['day_type', 'day_part', 'data_period', 'origin_subareaid','vehicle_weight'], columns="destination_subareaid",)
dfTruckDataBySubareaID_Pivot_ODTripVolume = dfTruckDataBySubareaID_Pivot_ODTripVolume.fillna(0)
dfTruckDataBySubareaID_Pivot_ODTripVolume = dfTruckDataBySubareaID_Pivot_ODTripVolume.reset_index()
dfTruckDataBySubareaID_Pivot_ODTripVolume

# Large District

dfTruckDataByDISTLRG = dfTruckDataByDISTLRG.fillna(-1)
display(dfTruckDataByDISTLRG.groupby(['origin_distlrg','vehicle_weight'],as_index=False).agg(od_trip_counts_sum=('od_trip_counts_sum','sum')))

dfTruckDataByDISTLRG_Pivot_NumRecords = dfTruckDataByDISTLRG.pivot_table(values="num_records",index=['day_type', 'day_part', 'data_period', 'origin_distlrg','vehicle_weight'], columns="destination_distlrg",)
dfTruckDataByDISTLRG_Pivot_NumRecords = dfTruckDataByDISTLRG_Pivot_NumRecords.fillna(0)
dfTruckDataByDISTLRG_Pivot_NumRecords = dfTruckDataByDISTLRG_Pivot_NumRecords.reset_index()
dfTruckDataByDISTLRG_Pivot_NumRecords

dfTruckDataByDISTLRG_Pivot_ODTripVolume = dfTruckDataByDISTLRG.pivot_table(values="od_trip_volume_sum",index=['day_type', 'day_part', 'data_period', 'origin_distlrg','vehicle_weight'], columns="destination_distlrg",)
dfTruckDataByDISTLRG_Pivot_ODTripVolume = dfTruckDataByDISTLRG_Pivot_ODTripVolume.fillna(0)
dfTruckDataByDISTLRG_Pivot_ODTripVolume = dfTruckDataByDISTLRG_Pivot_ODTripVolume.reset_index()
dfTruckDataByDISTLRG_Pivot_ODTripVolume


# Export to CSV

dfTruckDataByDISTLRG_Pivot_ODTripVolume.to_csv(os.path.join(results_folder,r"dfTruckDataByDISTLRG_Pivot.csv"))
