==> 2-Visualize Data.ipynb <==
import pandas as pd
#import arcpy

from arcgis.gis import *
gis = GIS()

#directories
working_directory = os.getcwd()
data_folder         = os.path.join(working_directory, "data"        )
intermediate_folder = os.path.join(working_directory, "intermediate")
results_folder      = os.path.join(working_directory, "results"     )

subareas_filename = os.path.join(data_folder, r"Subarea_2021_09_22.shp")
print(subareas_filename)

sdf_subareas = pd.DataFrame.spatial.from_featureclass(subareas_filename)
sdf_subareas

sdf_subareas

#create map1 with zoom set to Salt Lake (any city or place name can be used)
map_subareas = gis.map('Utah')

#add tract sdf
#sdf_subareas.spatial.plot(map_widget = map_subareas,
#                          renderer_type='u', # specify the unique value renderer using its notation 'u'
#                          col='SUBAREAID'  # column to get unique values from
#                         )

map_subareas.layout.height='600px'
map_subareas.legend = True

#display the map
map_subareas
